单元测试OK

+++ software_dzc-master/总体项目/后端/OAASystem/OAASystem/src/main/java/edu/neu/oaas/service/TenantService.java
@@ -168,6 +168,9 @@ public class TenantService {
 
     public boolean delete(Integer id) {
         Tenant tenant = tenantMapper.getById(id);
+        if (tenant == null) {
+            return false;
+        }
         departmentService.delete(tenant.getRootDepartmentId());
         tenantMapper.deleteById(id);
         return true;
